chore: move tests into separate files
Also chage coverage gathering mechanism again... Sorry for the churn.
