--- src/integrations/iam.rs
@@ -1,1 +1,0 @@
-pub struct IamResolver {}
